Split rejected application choice presenter into sub-implementations
There are 3 ways that reasons for rejecting an application can be stored, these each need their own presentation logic.
This commit adds three distinct presenters to handle the different formats
